requirements.txt
appdirs==1.4.4
ariadne==0.13.0
atomicwrites==1.4.0
attrs==21.2.0
black==21.7b0
blinker==1.4
click==8.0.1
click-completion==0.5.2
colorama==0.4.4
crayons==0.4.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
gitdb==4.0.7
GitPython==3.1.24
graphql-core==3.1.5
greenlet==1.1.1
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
py==1.10.0
pycycle==0.0.8
PyJWT==2.1.0
PyMySQL==1.0.2
pyotp==2.6.0
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.0
regex==2021.8.3
shellingham==1.4.0
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.22
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
